refactor: state management and platformsData
