+++ src/components/community/Platforms/types.ts
@@ -1,3 +1,5 @@
+import type { Platforms as PlatformsType } from "temporaryData/types"
+
 type Error = { code: number; message: string } | null
 type State = "initial" | "loading" | "success" | Error
 type ModalProps = {
@@ -11,14 +13,17 @@ type SignErrorProps = { error: Error }
 type JoinOrLeavePlatformProps = {
   platform: string
   communityId: number
-  title: string
-  description: string
-  buttonText?: string
+  isOpen: boolean
+  onClose: () => void
 }
 type PlatformButtonProps = {
   communityId: number
   isMember: boolean
   platform: string
+}
+type PlatformsProps = {
+  data: PlatformsType
+  community: number
 }
 
 export type {
@@ -28,4 +33,5 @@ export type {
   SignErrorProps,
   JoinOrLeavePlatformProps,
   PlatformButtonProps,
+  PlatformsProps,
 }
